fastapi_app/requirements.txt
uvicorn
fastapi
python-dotenv
groq